Simplify driver creation
