+++ operators/pkg/controller/elasticsearch/elasticsearch_controller.go
@@ -5,6 +5,7 @@
 package elasticsearch
 
 import (
+	"fmt"
 	"sync/atomic"
 	"time"
 
@@ -21,6 +22,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
 	"sigs.k8s.io/controller-runtime/pkg/source"
+
+	esversion "github.com/elastic/cloud-on-k8s/operators/pkg/controller/elasticsearch/version"
 
 	elasticsearchv1alpha1 "github.com/elastic/cloud-on-k8s/operators/pkg/apis/elasticsearch/v1alpha1"
 	"github.com/elastic/cloud-on-k8s/operators/pkg/controller/common"
@@ -243,11 +246,6 @@ func (r *ReconcileElasticsearch) internalReconcile(
 		return results
 	}
 
-	ver, err := commonversion.Parse(es.Spec.Version)
-	if err != nil {
-		return results.WithError(err)
-	}
-
 	violations, err := validation.Validate(es)
 	if err != nil {
 		return results.WithError(err)
@@ -257,23 +255,27 @@ func (r *ReconcileElasticsearch) internalReconcile(
 		return results
 	}
 
-	driver, err := driver.NewDriver(driver.Options{
-		Client:   r.Client,
-		Scheme:   r.scheme,
-		Recorder: r.recorder,
-
-		Version: *ver,
-
-		Expectations:   r.expectations,
-		Observers:      r.esObservers,
-		DynamicWatches: r.dynamicWatches,
-		Parameters:     r.Parameters,
-	})
+	ver, err := commonversion.Parse(es.Spec.Version)
 	if err != nil {
 		return results.WithError(err)
 	}
+	supported := esversion.SupportedVersions(*ver)
+	if supported == nil {
+		return results.WithError(fmt.Errorf("unsupported version: %s", ver))
+	}
 
-	return driver.Reconcile(es, reconcileState)
+	return driver.NewDefaultDriver(driver.DefaultDriverParameters{
+		OperatorParameters: r.Parameters,
+		ES:                 es,
+		ReconcileState:     reconcileState,
+		Client:             r.Client,
+		Scheme:             r.scheme,
+		Recorder:           r.recorder,
+		Version:            *ver,
+		Expectations:       r.expectations,
+		Observers:          r.esObservers,
+		DynamicWatches:     r.dynamicWatches,
+	}).Reconcile()
 }
 
 func (r *ReconcileElasticsearch) updateStatus(
